+++ elements/layercontrol/src/components/layer-tools.js
@@ -307,6 +307,10 @@ export class EOxLayerControlLayerTools extends LitElement {
       display: flex;
       justify-content: center;
     }
+    eox-layercontrol-tools-items.tools-list button.icon,
+    eox-layercontrol-tools-items.tools-list .button.icon {
+      margin-left: -6px;
+    }
     eox-layercontrol-tools-items button.icon::before,
     eox-layercontrol-tools-items .button.icon::before {
       width: 16px;
@@ -341,8 +345,11 @@ export class EOxLayerControlLayerTools extends LitElement {
       content: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' fill='%23004170' viewBox='0 0 24 24'%3E%3Ctitle%3Edrag-horizontal-variant%3C/title%3E%3Cpath d='M21 11H3V9H21V11M21 13H3V15H21V13Z' /%3E%3C/svg%3E");
     }
     [slot=info-content],
-    [slot=opacity-content] {
-      padding: 12px 6px;
+    [slot=opacity-content],
+    [slot=config-content],
+    [slot=datetime-content],
+    [slot=legend-content] {
+      padding: 6px 0;
     }
     [slot=info-content] * {
       max-width: 100%;