+++ elements/timecontrol/stories/index.js
@@ -6,3 +6,4 @@ export { default as SliderStory } from "./slider"; // Slider story
 export { default as ProgrammaticTimeSelectionStory } from "./programmatic-time-selection"; // Programmatic time selection story
 export { default as DisabledPlayButtonStory } from "./disabled-play-button"; // Disabled play button story
 export { default as NoMapStory } from "./no-map"; // No map provided story
+export { default as FormatStory } from "./format"; // Date-Time Format story