+++ elements/timecontrol/stories/slider.js
@@ -21,6 +21,7 @@ export const Slider = {
       .controlProperty=${args.controlProperty}
       .controlValues=${args.controlValues}
       .slider=${args.slider}
+      style="margin-top: 8px"
     ></eox-timecontrol>
   `,
 };